pyproject.toml
[project]
name = "bmspace"
version = "2.2.0"
description = "PACE BMS data retrieval and MQTT publishing"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.1",
    "PyYAML>=6.0",
    "pyserial>=3.5",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/bmspace"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bmspace"]
